cart routes for adding updating deleting added

+++ controllers/cartControllers.js
@@ -19,7 +19,16 @@ const createCart = async (userId) => {
 
 const getCart = async (userId, cartId, shouldPopulate = false) => {
   try {
-    let query = await Cart.findOne({ user: userId, _id: cartId });
+    let filter = {};
+    if (userId) {
+      filter.user = userId;
+    }
+
+    if (cartId) {
+      filter._id = cartId;
+    }
+
+    let query = Cart.findOne(filter);
 
     if (shouldPopulate) {
       query = query.populate({
@@ -30,7 +39,7 @@ const getCart = async (userId, cartId, shouldPopulate = false) => {
       });
     }
 
-    const cart = await query.exec();
+    const cart = await query.lean();
     return cart;
   } catch (error) {
     throw new Error(error);
@@ -43,8 +52,8 @@ const getCart = async (userId, cartId, shouldPopulate = false) => {
 
 const getUserCart = async (req, res, next) => {
   try {
-    const userId = req.user._id;
-    const cartId = req.params.cartId;
+    const userId = req?.user?._id;
+    const cartId = req.params.id;
 
     const cart = await getCart(userId, cartId, true);
 
@@ -52,9 +61,11 @@ const getUserCart = async (req, res, next) => {
       throw customError(404, "Cart not found!");
     }
 
-    const totalItems = cart?.items.length;
-    const subtotal = cart?.items.reduce((acc, cartItem) => {
-      acc + cartItem.price;
+    const totalItems = cart?.items.reduce((acc, item) => {
+      return acc + item.quantity;
+    }, 0);
+    const subtotal = cart?.items.reduce((acc, item) => {
+      return acc + item.product.price * item.quantity;
     }, 0);
 
     const response = {
@@ -74,7 +85,11 @@ const getUserCart = async (req, res, next) => {
 //access public
 const addToCart = async (req, res, next) => {
   try {
-    const { productId, userId, cartId } = req.body;
+    const { productId, quantity, userId, cartId } = req.body;
+
+    if (!productId || !quantity || quantity < 1) {
+      throw customError(400, "Product id and valid quantity is required");
+    }
 
     let cart = (await getCart(userId, cartId)) || (await createCart(userId));
 
@@ -84,12 +99,13 @@ const addToCart = async (req, res, next) => {
     });
 
     if (itemExists) {
-      itemExists.quantity += 1;
+      itemExists.quantity += quantity;
       await itemExists.save();
     } else {
       const newCartItem = await CartItem.create({
+        cartId: cart._id,
         product: productId,
-        quantity: 1,
+        quantity: quantity || 1,
       });
 
       cart.items.push(newCartItem._id);
@@ -110,13 +126,33 @@ const updateCartItem = async (req, res, next) => {
     const cartItemId = req.params.id;
     const { quantity } = req.body;
 
+    console.log(quantity);
+
+    if (
+      quantity === "" ||
+      quantity === undefined ||
+      quantity === null ||
+      quantity < 0
+    ) {
+      throw customError(400, "Valid quantity is required");
+    }
+
     if (quantity === 0) {
+      const cartItem = await CartItem.findOne({ _id: cartItemId });
+
+      if (!cartItem) {
+        throw customError(404, "Cart item not found");
+      }
+
+      const cartId = cartItem.cartId;
+
       await CartItem.deleteOne({ _id: cartItemId });
+      await Cart.updateOne({ _id: cartId }, { $pull: { items: cartItemId } });
       return res.status(200).send({ message: "Cart item deleted" });
     }
 
-    const updatedCartItem = await CartItem.findByIdAndUpdate(
-      cartItemId,
+    const updatedCartItem = await CartItem.findOneAndUpdate(
+      { _id: cartItemId },
       { $set: { quantity } },
       { new: true }
     );
@@ -125,7 +161,7 @@ const updateCartItem = async (req, res, next) => {
       throw customError(404, "Cart item not found");
     }
 
-    res.status(200).send(updateCartItem);
+    res.status(200).send(updatedCartItem);
   } catch (error) {
     next(error);
   }
